I've made two major changes in this commit:
1.  I've updated the project's group ID and all package names from `com.csj.framework` to `io.zabbixplus.framework`. This involved refactoring Java source files across all modules (`core`, `plugin-api`, `example-plugin`, `database`, `main-ui`), updating Spring Boot `scanBasePackages`, and modifying the `ServiceLoader` configuration in `example-plugin`.

2.  I've restructured the `main-ui` module to support a Vue.js 3 frontend, replacing the previous Thymeleaf server-side rendering setup:
    *   Thymeleaf dependencies and configurations were removed.
    *   Spring Boot in `main-ui` is now configured to serve static assets and includes a controller (`VueAppController`) to forward routes to `index.html` for SPA client-side routing.
    *   I integrated Node.js build tooling using the `com.github.node-gradle.node` Gradle plugin. This includes tasks for `npm install`, `npm run build`, and copying built frontend assets to Spring Boot's static resources.
    *   A conceptual Vue.js 3 application shell has been created in `main-ui/src/main/frontend/src/`, including basic components (App.vue, HomeView.vue), Vue Router setup, and placeholder views for plugin content.

3.  The `plugin-api` (`UiPlugin.java`) and its implementation in `example-plugin` (`SimpleExamplePlugin.java`) have been adapted conceptually for Vue.js. Plugins now specify a Vue component name via `getVueComponentName()` as their UI entry point.

**Blocker:**
I've encountered a significant challenge: the `npm install` command is consistently timing out. This is preventing me from downloading frontend dependencies, building the Vue.js application, and therefore any further development or testing of the Vue.js frontend features, including dynamic loading of plugin UI components.

The Java backend refactoring (group ID, package names) and the setup for Vue.js integration (Gradle, conceptual Vue app) are complete. Further progress on the Vue.js UI implementation is blocked by my inability to build the frontend.
