--- core/src/main/java/com/csj/framework/core/CoreApplication.java
@@ -1,12 +1,0 @@
-package com.csj.framework.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoreApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CoreApplication.class, args);
-    }
-}
