--- core/src/main/java/com/csj/framework/core/config/WebConfig.java
@@ -1,19 +1,0 @@
-package com.csj.framework.core.config;
-
-import com.csj.framework.core.interceptor.RequestLoggingInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Autowired
-    private RequestLoggingInterceptor requestLoggingInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(requestLoggingInterceptor);
-    }
-}
