--- core/src/main/java/com/csj/framework/core/exception/ResourceNotFoundException.java
@@ -1,11 +1,0 @@
-package com.csj.framework.core.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
