--- core/src/main/java/com/csj/framework/core/interceptor/RequestLoggingInterceptor.java
@@ -1,40 +1,0 @@
-package com.csj.framework.core.interceptor;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.ModelAndView;
-
-@Component
-public class RequestLoggingInterceptor implements HandlerInterceptor {
-
-    private static final Logger logger = LoggerFactory.getLogger(RequestLoggingInterceptor.class);
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        long startTime = System.currentTimeMillis();
-        request.setAttribute("startTime", startTime);
-        logger.info("Request URL::{} Request Method::{} ", request.getRequestURL().toString(), request.getMethod());
-        return true;
-    }
-
-    @Override
-    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
-        // Can log response details here if needed, but be careful with sensitive data in response body
-        // logger.info("Response Status::{}", response.getStatus());
-    }
-
-    @Override
-    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
-        long startTime = (Long) request.getAttribute("startTime");
-        long endTime = System.currentTimeMillis();
-        long executeTime = endTime - startTime;
-        logger.info("Request URL::{} Status::{} Time Taken::{}ms", request.getRequestURL().toString(), response.getStatus(), executeTime);
-        if (ex != null) {
-            logger.error("Error during request processing: ", ex);
-        }
-    }
-}
