--- core/src/main/java/com/csj/framework/core/plugin/PluginService.java
@@ -1,117 +1,0 @@
-package com.csj.framework.core.plugin;
-
-import com.csj.framework.plugin.Plugin; // From plugin-api
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-// import org.springframework.context.ApplicationContext; // If passing context to plugins
-// import org.springframework.beans.factory.annotation.Autowired; // If passing context
-import org.springframework.stereotype.Service;
-
-import jakarta.annotation.PostConstruct; // For @PostConstruct
-import jakarta.annotation.PreDestroy;  // For @PreDestroy
-
-import java.io.File;
-import java.net.URL;
-import java.net.URLClassLoader;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.ServiceLoader;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.Map;
-
-@Service
-public class PluginService {
-    private static final Logger logger = LoggerFactory.getLogger(PluginService.class);
-    private final Map<String, Plugin> loadedPlugins = new ConcurrentHashMap<>();
-    private final List<URLClassLoader> pluginClassLoaders = new ArrayList<>();
-
-    @Value("${framework.plugin.directory:./plugins}") // Default to a 'plugins' subdirectory
-    private String pluginDirectoryPath;
-
-    // @Autowired
-    // private ApplicationContext applicationContext; // If plugins need context
-
-    @PostConstruct
-    public void loadPlugins() {
-        File pluginDir = new File(pluginDirectoryPath);
-        if (!pluginDir.exists() || !pluginDir.isDirectory()) {
-            logger.warn("Plugin directory '{}' does not exist or is not a directory. Creating it.", pluginDirectoryPath);
-            if (!pluginDir.mkdirs()) {
-                logger.error("Could not create plugin directory '{}'. No plugins will be loaded.", pluginDirectoryPath);
-                return;
-            }
-        }
-
-        File[] pluginFiles = pluginDir.listFiles((dir, name) -> name.endsWith(".jar"));
-
-        if (pluginFiles == null || pluginFiles.length == 0) {
-            logger.info("No plugins found in directory '{}'.", pluginDirectoryPath);
-            return;
-        }
-
-        for (File pluginFile : pluginFiles) {
-            try {
-                logger.info("Attempting to load plugin from: {}", pluginFile.getAbsolutePath());
-                URL pluginUrl = pluginFile.toURI().toURL();
-                // Create a new classloader for each plugin JAR for better isolation,
-                // though true isolation is much harder (requires parent last or custom classloaders).
-                // For ServiceLoader to work across plugins loaded this way, the Plugin API
-                // must be accessible from the parent classloader (application's classloader).
-                URLClassLoader pluginClassLoader = new URLClassLoader(new URL[]{pluginUrl}, getClass().getClassLoader());
-                pluginClassLoaders.add(pluginClassLoader); // Keep track to close it later
-
-                // Using ServiceLoader to find implementations of the Plugin interface
-                ServiceLoader<Plugin> serviceLoader = ServiceLoader.load(Plugin.class, pluginClassLoader);
-                for (Plugin plugin : serviceLoader) {
-                    if (loadedPlugins.containsKey(plugin.getName())) {
-                        logger.warn("Plugin with name '{}' already loaded. Skipping duplicate from {}.", plugin.getName(), pluginFile.getName());
-                        continue;
-                    }
-                    // plugin.initialize(applicationContext); // If initializing with context
-                    plugin.load();
-                    loadedPlugins.put(plugin.getName(), plugin);
-                    logger.info("Successfully loaded plugin: {} - {}", plugin.getName(), plugin.getDescription());
-                }
-            } catch (Exception e) {
-                logger.error("Failed to load plugin from file: " + pluginFile.getName(), e);
-            }
-        }
-    }
-
-    public Map<String, Plugin> getLoadedPlugins() {
-        return new ConcurrentHashMap<>(loadedPlugins); // Return a copy
-    }
-
-    public Plugin getPlugin(String name) {
-        return loadedPlugins.get(name);
-    }
-
-    public String getPluginDirectoryPath() {
-        return pluginDirectoryPath;
-    }
-
-    @PreDestroy
-    public void unloadPlugins() {
-        logger.info("Unloading all plugins...");
-        loadedPlugins.forEach((name, plugin) -> {
-            try {
-                plugin.unload();
-                logger.info("Unloaded plugin: {}", name);
-            } catch (Exception e) {
-                logger.error("Error unloading plugin: " + name, e);
-            }
-        });
-        loadedPlugins.clear();
-        for (URLClassLoader classLoader : pluginClassLoaders) {
-            try {
-                classLoader.close();
-                logger.info("Closed plugin classloader: {}", classLoader);
-            } catch (Exception e) {
-                logger.error("Error closing plugin classloader", e);
-            }
-        }
-        pluginClassLoaders.clear();
-        logger.info("All plugins unloaded and classloaders closed.");
-    }
-}
