+++ core/src/main/java/io/zabbixplus/framework/core/CoreApplication.java
@@ -1,0 +1,12 @@
+package io.zabbixplus.framework.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"io.zabbixplus.framework.core"})
+public class CoreApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CoreApplication.class, args);
+    }
+}
