+++ core/src/main/java/io/zabbixplus/framework/core/config/WebConfig.java
@@ -1,0 +1,19 @@
+package io.zabbixplus.framework.core.config;
+
+import io.zabbixplus.framework.core.interceptor.RequestLoggingInterceptor; // Updated import
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Autowired
+    private RequestLoggingInterceptor requestLoggingInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(requestLoggingInterceptor);
+    }
+}
