+++ core/src/main/java/io/zabbixplus/framework/core/exception/ErrorResponse.java
@@ -1,0 +1,66 @@
+package io.zabbixplus.framework.core.exception; // Updated package
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class ErrorResponse {
+    private int status;
+    private String message;
+    private LocalDateTime timestamp;
+    private String path;
+    private List<String> details;
+
+    // Constructors
+    public ErrorResponse(int status, String message, String path) {
+        this.status = status;
+        this.message = message;
+        this.path = path;
+        this.timestamp = LocalDateTime.now();
+    }
+
+    public ErrorResponse(int status, String message, String path, List<String> details) {
+        this(status, message, path);
+        this.details = details;
+    }
+
+    // Getters and Setters
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(LocalDateTime timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+
+    public List<String> getDetails() {
+        return details;
+    }
+
+    public void setDetails(List<String> details) {
+        this.details = details;
+    }
+}
