+++ core/src/main/java/io/zabbixplus/framework/core/exception/ResourceNotFoundException.java
@@ -1,0 +1,11 @@
+package io.zabbixplus.framework.core.exception; // Updated package
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
