+++ core/src/main/java/io/zabbixplus/framework/core/web/ApiResponse.java
@@ -1,0 +1,62 @@
+package io.zabbixplus.framework.core.web; // Updated package
+
+import java.time.LocalDateTime;
+
+public class ApiResponse<T> {
+    private int status;
+    private String message;
+    private T data;
+    private LocalDateTime timestamp;
+
+    public ApiResponse(int status, String message, T data) {
+        this.status = status;
+        this.message = message;
+        this.data = data;
+        this.timestamp = LocalDateTime.now();
+    }
+
+    public static <T> ApiResponse<T> success(T data) {
+        return new ApiResponse<>(200, "Success", data);
+    }
+
+    public static <T> ApiResponse<T> success(String message, T data) {
+        return new ApiResponse<>(200, message, data);
+    }
+
+    public static ApiResponse<Void> success() {
+        return new ApiResponse<>(200, "Success", null);
+    }
+
+    // Getters and Setters
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public T getData() {
+        return data;
+    }
+
+    public void setData(T data) {
+        this.data = data;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(LocalDateTime timestamp) {
+        this.timestamp = timestamp;
+    }
+}
