--- example-plugin/src/main/resources/META-INF/services/com.csj.framework.plugin.Plugin
@@ -1,1 +1,0 @@
-com.csj.framework.exampleplugin.SimpleExamplePlugin
