+++ example-plugin/src/main/resources/META-INF/services/io.zabbixplus.framework.plugin.Plugin
@@ -1,0 +1,1 @@
+io.zabbixplus.framework.exampleplugin.SimpleExamplePlugin
