+++ main-ui/src/main/frontend/src/App.vue
@@ -1,0 +1,113 @@
+<template>
+  <div id="app-vue-main-container">
+    <nav class="main-navigation">
+      <router-link to="/">Home</router-link> |
+      <!-- Dynamic plugin links will be added here based on PluginService -->
+      <span v-for="pluginNav in pluginNavItems" :key="pluginNav.path">
+        <router-link :to="pluginNav.path">{{ pluginNav.name }}</router-link> |
+      </span>
+      <router-link to="/ui/plugins/list-vue">Loaded Plugins (Vue)</router-link>
+    </nav>
+    <main class="content-area">
+      <router-view/> <!-- This is where routed components will be displayed -->
+    </main>
+    <footer>
+      <p>&copy; Zabbix Plus Framework - Vue UI</p>
+    </footer>
+  </div>
+</template>
+
+<script>
+// In a real scenario, pluginNavItems would be fetched or managed by a global state/service
+// that interacts with the backend's PluginService.
+// For this conceptual step, it can be empty or have placeholders.
+import { ref, onMounted } from 'vue';
+
+export default {
+  name: 'App',
+  setup() {
+    const pluginNavItems = ref([]); // Placeholder for dynamic plugin navigation
+
+    // This is conceptual: how plugin nav items might be loaded.
+    // In reality, this needs an API call to the backend (e.g., to an endpoint exposing PluginService data).
+    const fetchPluginNavItems = async () => {
+      try {
+        // const response = await fetch('/api/ui/navigation-items'); // Example API endpoint
+        // if (!response.ok) throw new Error('Failed to fetch nav items');
+        // const data = await response.json();
+        // pluginNavItems.value = data; // Assuming data is an array of { name: '...', path: '...' }
+
+        // Placeholder data for conceptual development:
+        // This would ideally come from an API that gets data from PluginService -> UiPlugin.getNavigationItems()
+        // The paths should match those generated by UiPlugin implementations.
+        // Example: pluginNavItems.value = [
+        //   { name: "Example Plugin (Vue)", path: "/ui/plugin/SimpleExamplePlugin" }
+        // ];
+        console.log("Conceptual: fetchPluginNavItems called. In a real app, this would fetch from backend.");
+      } catch (error) {
+        console.error("Error fetching plugin navigation items:", error);
+        // Use placeholder/default items on error
+        // pluginNavItems.value = [
+        //     { name: "Error Loading Plugins", path: "/" }
+        // ];
+      }
+    };
+
+    onMounted(() => {
+      fetchPluginNavItems();
+    });
+
+    return {
+      pluginNavItems
+    };
+  }
+}
+</script>
+
+<style>
+#app-vue-main-container {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  /* text-align: center; */ /* Commented out to allow more standard layout */
+  color: #2c3e50;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.main-navigation {
+  padding: 15px;
+  background-color: #f8f9fa;
+  border-bottom: 1px solid #dee2e6;
+  text-align: center;
+  flex-shrink: 0;
+}
+
+.main-navigation a {
+  font-weight: bold;
+  color: #2c3e50;
+  margin: 0 10px;
+  text-decoration: none;
+}
+
+.main-navigation a.router-link-exact-active {
+  color: #42b983;
+}
+
+.content-area {
+  padding: 20px;
+  flex-grow: 1;
+}
+
+footer {
+  text-align: center;
+  padding: 10px;
+  /* margin-top: 20px; */ /* Removed for sticky footer effect */
+  font-size: 0.9em;
+  color: #6c757d;
+  border-top: 1px solid #eee;
+  background-color: #f8f9fa;
+  flex-shrink: 0;
+}
+</style>
