+++ main-ui/src/main/frontend/src/views/HomeView.vue
@@ -1,0 +1,21 @@
+<template>
+  <div class="home-view">
+    <h2>Home</h2>
+    <p>Welcome to the Zabbix Plus Framework main application!</p>
+    <p>This UI is built with Vue.js 3.</p>
+    <p>Navigation items for loaded UI plugins should appear in the header above if any plugins are active and provide UI contributions.</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'HomeView'
+}
+</script>
+
+<style scoped>
+.home-view {
+  padding: 20px;
+  text-align: center;
+}
+</style>
