+++ main-ui/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port=8000
 spring.application.name=main-ui-service
 
-# Basic Thymeleaf settings (can be refined later)
-spring.thymeleaf.cache=false # Good for development
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
+# Thymeleaf settings removed
+# spring.thymeleaf.cache=false
+# spring.thymeleaf.prefix=classpath:/templates/
+# spring.thymeleaf.suffix=.html
