+++ main-ui/src/main/resources/static/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Zabbix Plus Framework (Vue.js)</title>
+    <!-- Links to CSS/JS bundles will be added by Vue build process -->
+</head>
+<body>
+    <noscript>
+      <strong>We're sorry but this application doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+    </noscript>
+    <div id="app"></div>
+    <!-- Built files will be auto injected here by Vue CLI/Vite -->
+</body>
+</html>
